Create countdown timer component with name and value props with props validationfor each time block
