+++ package.json
@@ -20,6 +20,7 @@
     "aws-amplify": "^6.0.18",
     "aws-lambda": "^1.0.7",
     "next": "14.0.4",
+    "prop-types": "^15.8.1",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.0.1"
